+++ src/wallet/rpcwallet.cpp
@@ -3102,7 +3102,7 @@ static UniValue loadwallet(const Config &config,
         "Loads a wallet from a wallet file or directory."
         "\nNote that all wallet command-line options used when starting "
         "bitcoind will be"
-        "\napplied to the new wallet (eg -zapwallettxes, rescan, etc).\n",
+        "\napplied to the new wallet (eg -rescan, etc).\n",
         {
             {"filename", RPCArg::Type::STR, RPCArg::Optional::NO,
              "The wallet directory or .dat file."},