+++ test/lint/check-doc.py
@@ -48,6 +48,8 @@ SET_FALSE_POSITIVE_UNDOCUMENTED = set([
     '-forcecompactdb',
     '-parkdeepreorg',
     '-automaticunparking',
+    # Removed arguments that now just print a helpful error message
+    '-zapwallettxes',
     # Remove after November 2020 upgrade
     '-axionactivationtime',
     '-replayprotectionactivationtime',