+++ doc/release-notes.md
@@ -5,3 +5,6 @@ Bitcoin ABC version 0.24.3 is now available from:
   <https://download.bitcoinabc.org/0.24.3/>
 
 This release includes the following features and fixes:
+ - The `-zapwallettxes` startup option has been removed and its functionality removed
+   from the wallet. This functionality has been superseded with the abandon transaction
+   feature.